requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
celery==5.3.0
redis==4.5.1
django-cors-headers==3.13.0
python-dotenv==0.21.1
django-celery-beat==2.8.1
django-celery-results==2.6.0
requests==2.32.3
flower==2.0.0
django-debug-toolbar==4.2.0
